requirements.txt
asgiref==3.2.10
Babel==2.8.0
backcall==0.2.0
boto3==1.14.23
botocore==1.17.23
certifi==2020.6.20
chardet==3.0.4
colorama==0.4.3
coreapi==2.3.3
coreschema==0.0.4
decorator==4.4.2
defusedxml==0.6.0
dj-database-url==0.5.0
Django==3.0.7
django-allauth==0.42.0
django-cors-headers==3.4.0
django-cors-middleware==1.5.0
django-countries==6.1.2
django-filter==2.3.0
django-guardian==2.3.0
django-heroku==0.3.1
django-phonenumber-field==4.0.0
django-rest-auth==0.9.5
django-rest-swagger==2.2.0
django-storages==1.9.1
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.4.0
docutils==0.15.2
drf-yasg==1.17.1
entrypoints==0.3
gunicorn==20.0.4
idna==2.9
inflection==0.5.0
itypes==1.2.0
jedi
Jinja2==2.11.2
jmespath==0.10.0
Markdown==3.2.2
MarkupSafe==1.1.1
oauthlib==3.1.0
openapi-codec==1.3.2
packaging==20.4
parso==0.7.0
phonenumbers==8.12.6
pickleshare==0.7.5
Pillow==7.2.0
prompt-toolkit==3.0.5
psycopg2==2.8.5
Pygments==2.6.1
PyJWT==1.7.1
pyparsing==2.4.7
python-dateutil==2.8.1
python-memcached==1.59
python3-openid==3.1.0
pytz==2020.1
pyzmq==19.0.1
requests==2.24.0
requests-oauthlib==1.3.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
s3transfer==0.3.3
simplejson==3.17.0
six==1.15.0
sqlparse==0.3.1
tornado==6.0.4
traitlets==4.3.3
uritemplate==3.0.1
urllib3==1.25.9
wcwidth
whitenoise==5.1.0
wincertstore==0.2
